Organize bindings and add uv workflow
