+++ .github/workflows/uv.yml
@@ -1,0 +1,35 @@
+name: CI
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Install Rust
+        uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          override: true
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.12'
+      - name: Install uv
+        run: pip install uv
+      - name: Install maturin
+        run: pip install maturin
+      - name: Build wheel
+        run: maturin build --manifest-path rust_core/Cargo.toml --release --features python --strip -q
+      - name: Install wheel
+        run: pip install rust_core/target/wheels/*.whl
+      - name: Run tests
+        run: |
+          uv pip install -r requirements.txt || true
+          python -m unittest orderbook_tests.py -v
+      - name: Run benchmark
+        run: python benchmark.py
